refactor(image): extract optional-dep image helpers (Shutterbug/OpenCV/Applitools)
Move Shutterbug, OpenCV, and Applitools Eyes logic out of ImageProcessingActions into
dedicated package-private helper classes so the main class carries no optional-dep imports.
Add reflection-based unit tests to assert no optional types leak back into field/method signatures.
